refactor(ui): Remove unused components and routes, streamline routing structure for collections and dashboard
