+++ ui/src/features/collections/data/mock-collections.ts
@@ -1,0 +1,134 @@
+import type { Collection } from './types'
+
+export const collections: Collection[] = [
+  {
+    id: '1',
+    name: "Anand's collection",
+    domain: 'anands-collection-4m6pbv',
+    status: 'active',
+    sources: ['makenotion', 'asana', 'linear'],
+    lastSynced: '2 min ago',
+    createdAt: '2 min ago',
+    entities: 1247,
+    size: 4531,
+  },
+  {
+    id: '2',
+    name: 'Marketing Assets',
+    domain: 'marketing-assets-8k2nxp',
+    status: 'active',
+    sources: ['google', 'dropbox'],
+    lastSynced: '5 min ago',
+    createdAt: '2 days ago',
+    entities: 892,
+    size: 65432,
+  },
+  {
+    id: '3',
+    name: 'Engineering Docs',
+    domain: 'engineering-docs-3j9mvz',
+    status: 'syncing',
+    sources: ['github', 'atlassian', 'makenotion'],
+    lastSynced: '1 hour ago',
+    createdAt: '1 day ago',
+    entities: 3421,
+    size: 216,
+  },
+  {
+    id: '4',
+    name: 'Customer Support',
+    domain: 'customer-support-7r4kls',
+    status: 'active',
+    sources: ['zendesk', 'slackhq'],
+    lastSynced: '10 min ago',
+    entities: 5672,
+    createdAt: '1 week ago',
+    size: 92487,
+  },
+  {
+    id: '5',
+    name: 'Sales Pipeline',
+    domain: 'sales-pipeline-2m8fht',
+    status: 'active',
+    sources: ['hubspot', 'salesforce'],
+    lastSynced: '15 min ago',
+    entities: 2103,
+    createdAt: '1 month ago',
+    size: 1024,
+  },
+  {
+    id: '6',
+    name: 'Product Roadmap',
+    domain: 'product-roadmap-9n3wqx',
+    status: 'paused',
+    sources: ['linear', 'atlassian'],
+    lastSynced: '2 days ago',
+    entities: 456,
+    createdAt: '2 months ago',
+    size: 3121,
+  },
+  {
+    id: '7',
+    name: 'HR Resources',
+    domain: 'hr-resources-5k7pmt',
+    status: 'active',
+    sources: ['google', 'microsoft'],
+    lastSynced: '30 min ago',
+    entities: 1834,
+    createdAt: '3 months ago',
+    size: 5432,
+  },
+  {
+    id: '8',
+    name: 'Financial Records',
+    domain: 'financial-records-1h6cvn',
+    status: 'active',
+    sources: ['stripe', 'airtable'],
+    lastSynced: '1 hour ago',
+    entities: 7891,
+    createdAt: '3 months ago',
+    size: 1024,
+  },
+  {
+    id: '9',
+    name: 'Research Data',
+    domain: 'research-data-4w2jbp',
+    status: 'syncing',
+    sources: ['postgresql', 'google'],
+    lastSynced: '3 hours ago',
+    entities: 12456,
+    createdAt: '3 months ago',
+    size: 345231,
+  },
+  {
+    id: '10',
+    name: 'Client Projects',
+    domain: 'client-projects-8m5rzq',
+    status: 'active',
+    sources: ['asana', 'box', 'slackhq'],
+    lastSynced: '45 min ago',
+    entities: 3287,
+    createdAt: '3 months ago',
+    size: 345231,
+  },
+]
+
+export const sourceSlugMap: Record<string, string> = {
+  notion: 'makenotion',
+  makenotion: 'makenotion',
+  asana: 'asana',
+  linear: 'linear',
+  google: 'google',
+  dropbox: 'dropbox',
+  github: 'github',
+  atlassian: 'atlassian',
+  zendesk: 'zendesk',
+  slackhq: 'slackhq',
+  hubspot: 'hubspot',
+  salesforce: 'salesforce',
+  microsoft: 'microsoft',
+  stripe: 'stripe',
+  airtable: 'airtable',
+  postgresql: 'postgresql',
+  box: 'box',
+}
