+++ ui/src/features/collections/data/types.ts
@@ -1,0 +1,11 @@
+export interface Collection {
+  id: string
+  name: string
+  domain: string
+  status: 'active' | 'syncing' | 'paused'
+  sources: string[]
+  lastSynced: string
+  createdAt: string
+  entities: number
+  size: number
+}
