+++ ui/src/features/components-page/sections/navigation.tsx
@@ -1,0 +1,295 @@
+import {
+  Breadcrumb,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbList,
+  BreadcrumbPage,
+  BreadcrumbSeparator,
+} from '@/components/ui/breadcrumb'
+import {
+  Command,
+  CommandEmpty,
+  CommandGroup,
+  CommandInput,
+  CommandItem,
+  CommandList,
+  CommandSeparator,
+} from '@/components/ui/command'
+import {
+  Menubar,
+  MenubarContent,
+  MenubarItem,
+  MenubarMenu,
+  MenubarSeparator,
+  MenubarShortcut,
+  MenubarTrigger,
+} from '@/components/ui/menubar'
+import {
+  NavigationMenu,
+  NavigationMenuContent,
+  NavigationMenuItem,
+  NavigationMenuLink,
+  NavigationMenuList,
+  NavigationMenuTrigger,
+} from '@/components/ui/navigation-menu'
+import {
+  Pagination,
+  PaginationContent,
+  PaginationEllipsis,
+  PaginationItem,
+  PaginationLink,
+  PaginationNext,
+  PaginationPrevious,
+} from '@/components/ui/pagination'
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
+import {
+  Calculator,
+  Calendar as CalendarIcon,
+  CreditCard,
+  Settings,
+  Smile,
+  User,
+} from 'lucide-react'
+
+export function BreadcrumbSection() {
+  return (
+    <section className="space-y-4">
+      <h2 className="text-xs font-mono uppercase tracking-widest text-muted-foreground">
+        Breadcrumb
+      </h2>
+      <Breadcrumb>
+        <BreadcrumbList>
+          <BreadcrumbItem>
+            <BreadcrumbLink href="#">Home</BreadcrumbLink>
+          </BreadcrumbItem>
+          <BreadcrumbSeparator />
+          <BreadcrumbItem>
+            <BreadcrumbLink href="#">Components</BreadcrumbLink>
+          </BreadcrumbItem>
+          <BreadcrumbSeparator />
+          <BreadcrumbItem>
+            <BreadcrumbPage>Breadcrumb</BreadcrumbPage>
+          </BreadcrumbItem>
+        </BreadcrumbList>
+      </Breadcrumb>
+    </section>
+  )
+}
+
+export function CommandSection() {
+  return (
+    <section className="space-y-4">
+      <h2 className="text-xs font-mono uppercase tracking-widest text-muted-foreground">
+        Command
+      </h2>
+      <Command className="rounded-lg border shadow-md max-w-md">
+        <CommandInput placeholder="Type a command or search..." />
+        <CommandList>
+          <CommandEmpty>No results found.</CommandEmpty>
+          <CommandGroup heading="Suggestions">
+            <CommandItem>
+              <CalendarIcon className="mr-2 size-4" />
+              <span>Calendar</span>
+            </CommandItem>
+            <CommandItem>
+              <Smile className="mr-2 size-4" />
+              <span>Search Emoji</span>
+            </CommandItem>
+            <CommandItem>
+              <Calculator className="mr-2 size-4" />
+              <span>Calculator</span>
+            </CommandItem>
+          </CommandGroup>
+          <CommandSeparator />
+          <CommandGroup heading="Settings">
+            <CommandItem>
+              <User className="mr-2 size-4" />
+              <span>Profile</span>
+            </CommandItem>
+            <CommandItem>
+              <CreditCard className="mr-2 size-4" />
+              <span>Billing</span>
+            </CommandItem>
+            <CommandItem>
+              <Settings className="mr-2 size-4" />
+              <span>Settings</span>
+            </CommandItem>
+          </CommandGroup>
+        </CommandList>
+      </Command>
+    </section>
+  )
+}
+
+export function MenubarSection() {
+  return (
+    <section className="space-y-4">
+      <h2 className="text-xs font-mono uppercase tracking-widest text-muted-foreground">
+        Menubar
+      </h2>
+      <Menubar>
+        <MenubarMenu>
+          <MenubarTrigger>File</MenubarTrigger>
+          <MenubarContent>
+            <MenubarItem>
+              New Tab <MenubarShortcut>⌘T</MenubarShortcut>
+            </MenubarItem>
+            <MenubarItem>
+              New Window <MenubarShortcut>⌘N</MenubarShortcut>
+            </MenubarItem>
+            <MenubarSeparator />
+            <MenubarItem>Share</MenubarItem>
+            <MenubarSeparator />
+            <MenubarItem>Print</MenubarItem>
+          </MenubarContent>
+        </MenubarMenu>
+        <MenubarMenu>
+          <MenubarTrigger>Edit</MenubarTrigger>
+          <MenubarContent>
+            <MenubarItem>
+              Undo <MenubarShortcut>⌘Z</MenubarShortcut>
+            </MenubarItem>
+            <MenubarItem>
+              Redo <MenubarShortcut>⇧⌘Z</MenubarShortcut>
+            </MenubarItem>
+            <MenubarSeparator />
+            <MenubarItem>Cut</MenubarItem>
+            <MenubarItem>Copy</MenubarItem>
+            <MenubarItem>Paste</MenubarItem>
+          </MenubarContent>
+        </MenubarMenu>
+        <MenubarMenu>
+          <MenubarTrigger>View</MenubarTrigger>
+          <MenubarContent>
+            <MenubarItem>Reload</MenubarItem>
+            <MenubarSeparator />
+            <MenubarItem>Toggle Fullscreen</MenubarItem>
+          </MenubarContent>
+        </MenubarMenu>
+      </Menubar>
+    </section>
+  )
+}
+
+export function NavigationMenuSection() {
+  return (
+    <section className="space-y-4">
+      <h2 className="text-xs font-mono uppercase tracking-widest text-muted-foreground">
+        Navigation Menu
+      </h2>
+      <NavigationMenu>
+        <NavigationMenuList>
+          <NavigationMenuItem>
+            <NavigationMenuTrigger>Getting Started</NavigationMenuTrigger>
+            <NavigationMenuContent>
+              <ul className="grid gap-3 p-4 md:w-[400px] lg:w-[500px] lg:grid-cols-[.75fr_1fr]">
+                <li className="row-span-3">
+                  <NavigationMenuLink className="flex h-full w-full select-none flex-col justify-end rounded-md bg-gradient-to-b from-muted/50 to-muted p-6 no-underline outline-none focus:shadow-md">
+                    <div className="mb-2 mt-4 text-lg font-medium">
+                      shadcn/ui
+                    </div>
+                    <p className="text-sm leading-tight text-muted-foreground">
+                      Beautifully designed components built with Radix UI
+                      and Tailwind CSS.
+                    </p>
+                  </NavigationMenuLink>
+                </li>
+                <li>
+                  <NavigationMenuLink>Introduction</NavigationMenuLink>
+                </li>
+                <li>
+                  <NavigationMenuLink>Installation</NavigationMenuLink>
+                </li>
+                <li>
+                  <NavigationMenuLink>Typography</NavigationMenuLink>
+                </li>
+              </ul>
+            </NavigationMenuContent>
+          </NavigationMenuItem>
+          <NavigationMenuItem>
+            <NavigationMenuTrigger>Components</NavigationMenuTrigger>
+            <NavigationMenuContent>
+              <ul className="grid w-[400px] gap-3 p-4 md:w-[500px] md:grid-cols-2 lg:w-[600px]">
+                <li>
+                  <NavigationMenuLink>Alert Dialog</NavigationMenuLink>
+                </li>
+                <li>
+                  <NavigationMenuLink>Hover Card</NavigationMenuLink>
+                </li>
+                <li>
+                  <NavigationMenuLink>Progress</NavigationMenuLink>
+                </li>
+                <li>
+                  <NavigationMenuLink>Scroll-area</NavigationMenuLink>
+                </li>
+              </ul>
+            </NavigationMenuContent>
+          </NavigationMenuItem>
+          <NavigationMenuItem>
+            <NavigationMenuLink>Documentation</NavigationMenuLink>
+          </NavigationMenuItem>
+        </NavigationMenuList>
+      </NavigationMenu>
+    </section>
+  )
+}
+
+export function PaginationSection() {
+  return (
+    <section className="space-y-4">
+      <h2 className="text-xs font-mono uppercase tracking-widest text-muted-foreground">
+        Pagination
+      </h2>
+      <Pagination>
+        <PaginationContent>
+          <PaginationItem>
+            <PaginationPrevious href="#" />
+          </PaginationItem>
+          <PaginationItem>
+            <PaginationLink href="#">1</PaginationLink>
+          </PaginationItem>
+          <PaginationItem>
+            <PaginationLink href="#" isActive>
+              2
+            </PaginationLink>
+          </PaginationItem>
+          <PaginationItem>
+            <PaginationLink href="#">3</PaginationLink>
+          </PaginationItem>
+          <PaginationItem>
+            <PaginationEllipsis />
+          </PaginationItem>
+          <PaginationItem>
+            <PaginationNext href="#" />
+          </PaginationItem>
+        </PaginationContent>
+      </Pagination>
+    </section>
+  )
+}
+
+export function TabsSection() {
+  return (
+    <section className="space-y-4">
+      <h2 className="text-xs font-mono uppercase tracking-widest text-muted-foreground">
+        Tabs
+      </h2>
+      <Tabs defaultValue="account" className="w-full max-w-md">
+        <TabsList>
+          <TabsTrigger value="account">Account</TabsTrigger>
+          <TabsTrigger value="password">Password</TabsTrigger>
+          <TabsTrigger value="settings">Settings</TabsTrigger>
+        </TabsList>
+        <TabsContent value="account">
+          Make changes to your account here.
+        </TabsContent>
+        <TabsContent value="password">
+          Change your password here.
+        </TabsContent>
+        <TabsContent value="settings">
+          Manage your settings here.
+        </TabsContent>
+      </Tabs>
+    </section>
+  )
+}
