+++ ui/src/features/source-connections/components/status-badge.tsx
@@ -1,0 +1,58 @@
+import { Badge } from '@/components/ui/badge'
+import {
+  IconAlertCircle,
+  IconClock,
+  IconKey,
+  IconLoader2,
+} from '@tabler/icons-react'
+import type { SourceConnection } from '../data/types'
+
+export function StatusBadge({ status }: { status: SourceConnection['status'] }) {
+  const statusConfig = {
+    active: {
+      indicator: <span className="size-1.5 rounded-full bg-emerald-500" />,
+      badge:
+        'bg-emerald-50 border-emerald-200 text-emerald-700 dark:bg-emerald-950 dark:border-emerald-800 dark:text-emerald-400',
+      label: 'Active',
+    },
+    syncing: {
+      indicator: <IconLoader2 className="size-3 animate-spin" />,
+      badge:
+        'bg-amber-50 border-amber-200 text-amber-700 dark:bg-amber-950 dark:border-amber-800 dark:text-amber-400',
+      label: 'Syncing',
+    },
+    pending_auth: {
+      indicator: <IconKey className="size-3" />,
+      badge:
+        'bg-blue-50 border-blue-200 text-blue-700 dark:bg-blue-950 dark:border-blue-800 dark:text-blue-400',
+      label: 'Pending Auth',
+    },
+    pending_sync: {
+      indicator: <IconClock className="size-3" />,
+      badge:
+        'bg-gray-50 border-gray-200 text-gray-600 dark:bg-gray-900 dark:border-gray-700 dark:text-gray-400',
+      label: 'Pending Sync',
+    },
+    error: {
+      indicator: <IconAlertCircle className="size-3" />,
+      badge:
+        'bg-red-50 border-red-200 text-red-700 dark:bg-red-950 dark:border-red-800 dark:text-red-400',
+      label: 'Error',
+    },
+    inactive: {
+      indicator: <span className="size-1.5 rounded-full bg-gray-400" />,
+      badge:
+        'bg-gray-50 border-gray-200 text-gray-600 dark:bg-gray-900 dark:border-gray-700 dark:text-gray-400',
+      label: 'Inactive',
+    },
+  }
+
+  const config = statusConfig[status]
+
+  return (
+    <Badge className={`${config.badge} gap-1.5`}>
+      {config.indicator}
+      {config.label}
+    </Badge>
+  )
+}
