+++ ui/src/features/source-connections/data/mock-connections.ts
@@ -1,0 +1,310 @@
+import type { SourceConnection } from './types'
+
+export const sourceConnections: SourceConnection[] = [
+  {
+    id: '1',
+    account: 'anand@airweave.ai',
+    shortName: 'gmail',
+    collectionId: 'anands-collection-4m6pbv',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 1247,
+    federatedSearch: false,
+    createdAt: '2 days ago',
+    modifiedAt: '5 min ago',
+  },
+  {
+    id: '2',
+    account: 'U04MDQX8R7P',
+    shortName: 'slack',
+    collectionId: 'marketing-assets-8k2nxp',
+    status: 'syncing',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 892,
+    federatedSearch: false,
+    createdAt: '1 week ago',
+    modifiedAt: '2 min ago',
+  },
+  {
+    id: '3',
+    account: 'anand@company.notion.so',
+    shortName: 'notion',
+    collectionId: 'engineering-docs-3j9mvz',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 3421,
+    federatedSearch: true,
+    createdAt: '2 weeks ago',
+    modifiedAt: '1 hour ago',
+  },
+  {
+    id: '4',
+    account: 'AnandChowdhary',
+    shortName: 'github',
+    collectionId: 'engineering-docs-3j9mvz',
+    status: 'pending_auth',
+    authMethod: 'oauth_browser',
+    isAuthenticated: false,
+    entityCount: 0,
+    federatedSearch: false,
+    createdAt: '1 day ago',
+    modifiedAt: '1 day ago',
+  },
+  {
+    id: '5',
+    account: '0053h000002xQ1YAAU',
+    shortName: 'salesforce',
+    collectionId: 'sales-pipeline-2m8fht',
+    status: 'error',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 2103,
+    federatedSearch: false,
+    createdAt: '1 month ago',
+    modifiedAt: '3 hours ago',
+  },
+  {
+    id: '6',
+    account: 'anand@airweave.atlassian.net',
+    shortName: 'atlassian',
+    collectionId: 'product-roadmap-9n3wqx',
+    status: 'inactive',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 456,
+    federatedSearch: false,
+    createdAt: '2 months ago',
+    modifiedAt: '2 weeks ago',
+  },
+  {
+    id: '7',
+    account: 'anand@airweave.ai',
+    shortName: 'google',
+    collectionId: 'hr-resources-5k7pmt',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 1834,
+    federatedSearch: true,
+    createdAt: '3 months ago',
+    modifiedAt: '30 min ago',
+  },
+  {
+    id: '8',
+    account: 'acct_1NqX2kLkdIwHu7ix',
+    shortName: 'stripe',
+    collectionId: 'financial-records-1h6cvn',
+    status: 'pending_sync',
+    authMethod: 'direct',
+    isAuthenticated: true,
+    entityCount: 0,
+    federatedSearch: false,
+    createdAt: '1 hour ago',
+    modifiedAt: '1 hour ago',
+  },
+  {
+    id: '9',
+    account: 'postgres@db.airweave.ai',
+    shortName: 'postgresql',
+    collectionId: 'research-data-4w2jbp',
+    status: 'active',
+    authMethod: 'direct',
+    isAuthenticated: true,
+    entityCount: 12456,
+    federatedSearch: false,
+    createdAt: '3 months ago',
+    modifiedAt: '15 min ago',
+  },
+  {
+    id: '10',
+    account: '1204823456789012',
+    shortName: 'asana',
+    collectionId: 'client-projects-8m5rzq',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 3287,
+    federatedSearch: false,
+    createdAt: '3 months ago',
+    modifiedAt: '45 min ago',
+  },
+  {
+    id: '11',
+    account: 'sales@airweave.ai',
+    shortName: 'hubspot',
+    collectionId: 'customer-success-9k2lxy',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 2789,
+    federatedSearch: true,
+    createdAt: '4 days ago',
+    modifiedAt: '20 min ago',
+  },
+  {
+    id: '12',
+    account: 'support@airweave.ai',
+    shortName: 'zendesk',
+    collectionId: 'support-tickets-2b4cfd',
+    status: 'error',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 1543,
+    federatedSearch: false,
+    createdAt: '1 week ago',
+    modifiedAt: '10 min ago',
+  },
+  {
+    id: '13',
+    account: 'files@airweave.ai',
+    shortName: 'dropbox',
+    collectionId: 'design-assets-7m2kxr',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 987,
+    federatedSearch: false,
+    createdAt: '2 weeks ago',
+    modifiedAt: '3 hours ago',
+  },
+  {
+    id: '14',
+    account: 'legal@airweave.ai',
+    shortName: 'box',
+    collectionId: 'contracts-1k9bqt',
+    status: 'inactive',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 312,
+    federatedSearch: false,
+    createdAt: '1 month ago',
+    modifiedAt: '1 week ago',
+  },
+  {
+    id: '15',
+    account: 'm365@airweave.ai',
+    shortName: 'microsoft',
+    collectionId: 'sharepoint-sites-5t8pvr',
+    status: 'pending_sync',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 0,
+    federatedSearch: true,
+    createdAt: '3 days ago',
+    modifiedAt: '3 days ago',
+  },
+  {
+    id: '16',
+    account: 'ops@airweave.ai',
+    shortName: 'airtable',
+    collectionId: 'ops-tracking-6y1nqd',
+    status: 'syncing',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 1456,
+    federatedSearch: false,
+    createdAt: '5 days ago',
+    modifiedAt: '1 min ago',
+  },
+  {
+    id: '17',
+    account: 'bugs@airweave.ai',
+    shortName: 'linear',
+    collectionId: 'issue-tracker-1r6fzp',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 2154,
+    federatedSearch: false,
+    createdAt: '2 months ago',
+    modifiedAt: '12 min ago',
+  },
+  {
+    id: '18',
+    account: 'datasync@airweave.ai',
+    shortName: 'postgresql',
+    collectionId: 'analytics-warehouse-3g7tmd',
+    status: 'pending_auth',
+    authMethod: 'direct',
+    isAuthenticated: false,
+    entityCount: 0,
+    federatedSearch: false,
+    createdAt: '30 min ago',
+    modifiedAt: '30 min ago',
+  },
+  {
+    id: '19',
+    account: 'anand@airweave.ai',
+    shortName: 'gmail',
+    collectionId: 'lead-intel-7b3kpv',
+    status: 'active',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 1678,
+    federatedSearch: false,
+    createdAt: '6 months ago',
+    modifiedAt: '5 min ago',
+  },
+  {
+    id: '20',
+    account: 'engineering@airweave.ai',
+    shortName: 'github',
+    collectionId: 'platform-repos-2f9cjh',
+    status: 'syncing',
+    authMethod: 'oauth_browser',
+    isAuthenticated: true,
+    entityCount: 4321,
+    federatedSearch: true,
+    createdAt: '4 months ago',
+    modifiedAt: 'just now',
+  },
+]
+
+export const collectionNameMap: Record<string, string> = {
+  'anands-collection-4m6pbv': "Anand's collection",
+  'marketing-assets-8k2nxp': 'Marketing assets',
+  'engineering-docs-3j9mvz': 'Engineering docs',
+  'sales-pipeline-2m8fht': 'Sales pipeline',
+  'product-roadmap-9n3wqx': 'Product roadmap',
+  'hr-resources-5k7pmt': 'HR resources',
+  'financial-records-1h6cvn': 'Financial records',
+  'research-data-4w2jbp': 'Research data',
+  'client-projects-8m5rzq': 'Client projects',
+  'customer-success-9k2lxy': 'Customer success',
+  'support-tickets-2b4cfd': 'Support tickets',
+  'design-assets-7m2kxr': 'Design assets',
+  'contracts-1k9bqt': 'Contracts',
+  'sharepoint-sites-5t8pvr': 'SharePoint sites',
+  'ops-tracking-6y1nqd': 'Ops tracking',
+  'issue-tracker-1r6fzp': 'Issue tracker',
+  'analytics-warehouse-3g7tmd': 'Analytics warehouse',
+  'lead-intel-7b3kpv': 'Lead intelligence',
+  'platform-repos-2f9cjh': 'Platform repos',
+}
+
+export const sourceSlugMap: Record<string, string> = {
+  gmail: 'google',
+  slack: 'slackhq',
+  notion: 'makenotion',
+  github: 'github',
+  salesforce: 'salesforce',
+  atlassian: 'atlassian',
+  google: 'google',
+  stripe: 'stripe',
+  postgresql: 'postgresql',
+  asana: 'asana',
+  hubspot: 'hubspot',
+  zendesk: 'zendesk',
+  dropbox: 'dropbox',
+  box: 'box',
+  microsoft: 'microsoft',
+  airtable: 'airtable',
+  linear: 'linear',
+}
+
+export function getCollectionName(id: string) {
+  return collectionNameMap[id] ?? id
+}
